Retrieve Coefficients for specific state
"READ_DATCOM.m" is a function that takes as input the current "alpha-beta-Mach-altitude" state and provides the aerodynamic coefficients. It shall be included in the simulator.

The script "Auto_Matrices_sim.m" is called inside the previous function, and it shall be modified in order to create the apropriate "for005.dat", evaluate it with DATCOM, and provide the corresponding "for006.dat" in order to retrieve the coefficients.